requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
alembic
python-dotenv
streamlit
